refactor: move lifecycle logic to per-indexer container
